+++ src/App.tsx
@@ -1,14 +1,76 @@
 import * as React from 'react';
 import './i18n';
+import AppAppBar from './app-appbar';
 import SVGs from './svgs';
 // import SVGs from './svgs';
 import Panels from './panels';
 import { getParams } from './utils';
 import { getBranches, useTpo, getRoutes } from './methods';
-import { ParamContext, paramReducer } from './context';
+import { CanvasContext, ParamContext, paramReducer } from './context';
+import { createMuiTheme, ThemeProvider, useMediaQuery, LinearProgress } from '@material-ui/core';
+
+const darkTheme = createMuiTheme({
+    palette: {
+        type: 'dark',
+        primary: {
+            main: '#bb86fc',
+        },
+        secondary: {
+            main: '#018786',
+        },
+        error: {
+            main: '#cf6679',
+        },
+    },
+    overrides: {
+        MuiDialog: {
+            paper: {
+                minWidth: 280,
+            },
+        },
+    },
+});
+
+const lightTheme = createMuiTheme({
+    palette: {
+        primary: {
+            light: '#bb86fc',
+            main: '#6200ee',
+        },
+        secondary: {
+            main: '#018786',
+        },
+        error: {
+            main: '#b00020',
+        },
+    },
+    overrides: {
+        MuiDialog: {
+            paper: {
+                minWidth: 280,
+            },
+        },
+    },
+});
 
 export default function App() {
-    // const [param, setParam] = React.useState(getParams());
+    const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
+    const theme = prefersDarkMode ? darkTheme : lightTheme;
+
+    const canvasAvailable = ((style): ('destination' | 'runin' | 'railmap')[] => {
+        switch (style) {
+            case 'mtr':
+                return ['destination', 'railmap'];
+            case 'gzmtr':
+                return ['runin', 'railmap'];
+            case 'shmetro':
+                return ['destination', 'runin', 'railmap'];
+            default:
+                return [];
+        }
+    })(window.urlParams.get('style'));
+    const [canvasToShown, setCanvasToShown] = React.useState<'all' | 'destination' | 'runin' | 'railmap'>('all');
+
     const [param, dispatch] = React.useReducer(paramReducer, getParams());
     const paramString = JSON.stringify(param);
     React.useEffect(() => localStorage.setItem('rmgParam', paramString), [paramString]);
@@ -32,19 +94,27 @@ export default function App() {
     const handleUpdate = (key, data) => dispatch({ type: 'ANY', key, data });
 
     return (
-        // <React.Suspense fallback="loading">
-        <ParamContext.Provider
-            value={{
-                param,
-                dispatch,
-                branches,
-                routes,
-                deps,
-            }}
-        >
-            <SVGs />
-            <Panels param={param} paramUpdate={handleUpdate} tpo={tpo} />
-        </ParamContext.Provider>
-        // </React.Suspense>
+        <>
+            <ThemeProvider theme={theme}>
+                <CanvasContext.Provider value={{ canvasAvailable, canvasToShown, setCanvasToShown }}>
+                    <React.Suspense fallback={<LinearProgress />}>
+                        <AppAppBar />
+                    </React.Suspense>
+                    <ParamContext.Provider
+                        value={{
+                            param,
+                            dispatch,
+                            branches,
+                            routes,
+                            deps,
+                        }}
+                    >
+                        <SVGs />
+                        <Panels param={param} paramUpdate={handleUpdate} tpo={tpo} />
+                    </ParamContext.Provider>
+                </CanvasContext.Provider>
+                <canvas style={{ display: 'none' }} />
+            </ThemeProvider>
+        </>
     );
 }